custom view

+++ app/src/main/java/home/smart/fly/animationdemo/swipeanim/lib/ItemTouchListener.java
@@ -11,5 +11,5 @@ public interface ItemTouchListener {
 
     void show();
 
-    void hide(float dy);
+    void hide(int dy);
 }
